% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cv.knn}
\alias{cv.knn}
\title{Perform k-fold cross validation}
\usage{
cv.knn(dset, k = 10, nFolds = 10, nRuns = 10, alpha = 0.05)
}
\arguments{
\item{dset}{label ranking dataset}

\item{k}{the number of nearest neighbors to search. The default value is set to 10}

\item{nFolds}{the number of folds for the cross validation}

\item{nRuns}{the number of runs for the cross validation}

\item{alpha}{the default value is set to 0.05}
}
\value{
???
}
\description{
Perform k-fold cross validation
}
\examples{
# perform the 10 runs of 10-folds cross validation on the Iris dataset
dset = data(data.iris)
cv.knn(dset, k=15, nFolds=10, nRuns=10, alpha=0.05)
}

