% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{data.authorship}
\alias{data.authorship}
\title{Authorship dataset}
\format{A data frame with 841 observations, 70 attributes and 4 labels}
\source{
UCI respository and Statlog collection
}
\usage{
data(data.authorship)
}
\description{
Authorship dataset
}
\examples{
data(data.authorship)
}

