% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{data.bodyfat}
\alias{data.bodyfat}
\title{Bodyfat dataset}
\format{A data frame with 252 observations, 7 attributes and 7 labels}
\source{
UCI respository and Statlog collection
}
\usage{
data(data.bodyfat)
}
\description{
Bodyfat dataset
}
\examples{
data(data.bodyfat)
}

