% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{data.calhousing}
\alias{data.calhousing}
\title{Calhousing dataset}
\format{A data frame with 20640 observations, 4 attributes and 4 labels}
\source{
UCI respository and Statlog collection
}
\usage{
data(data.calhousing)
}
\description{
Calhousing dataset
}
\examples{
data(data.calhousing)
}

