% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{data.cpu}
\alias{data.cpu}
\title{cpu dataset}
\format{A data frame with 8192 observations, 6 attributes and 5 labels}
\source{
UCI respository and Statlog collection
}
\usage{
data(data.cpu)
}
\description{
cpu dataset
}
\examples{
data(data.cpu)
}

