% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{data.elevators}
\alias{data.elevators}
\title{elevators dataset}
\format{A data frame with 16599 observations, 9 attributes and 9 labels}
\source{
UCI respository and Statlog collection
}
\usage{
data(data.elevators)
}
\description{
elevators dataset
}
\examples{
data(data.elevators)
}

