% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{data.fried}
\alias{data.fried}
\title{fried dataset}
\format{A data frame with 40769 observations, 9 attributes and 5 labels}
\source{
UCI respository and Statlog collection
}
\usage{
data(data.fried)
}
\description{
fried dataset
}
\examples{
data(data.fried)
}

