% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{data.glass}
\alias{data.glass}
\title{glass dataset}
\format{A data frame with 214 observations, 9 attributes and 6 labels}
\source{
UCI respository and Statlog collection
}
\usage{
data(data.glass)
}
\description{
glass dataset
}
\examples{
data(data.glass)
}

