% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{data.housing}
\alias{data.housing}
\title{housing dataset}
\format{A data frame with 506 observations, 6 attributes and 6 labels}
\source{
UCI respository and Statlog collection
}
\usage{
data(data.housing)
}
\description{
housing dataset
}
\examples{
data(data.housing)
}

