% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{data.iris}
\alias{data.iris}
\title{iris dataset}
\format{A data frame with 150 observations, 4 attributes and 3 labels}
\source{
UCI respository and Statlog collection
}
\usage{
data(data.iris)
}
\description{
iris dataset
}
\examples{
data(data.iris)
}

