% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{data.pendigits}
\alias{data.pendigits}
\title{pendigits dataset}
\format{A data frame with 10992 observations, 16 attributes and 10 labels}
\source{
UCI respository and Statlog collection
}
\usage{
data(data.pendigits)
}
\description{
pendigits dataset
}
\examples{
data(data.pendigits)
}

