% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{data.segment}
\alias{data.segment}
\title{segment dataset}
\format{A data frame with 2310 observations, 18 attributes and 7 labels}
\source{
UCI respository and Statlog collection
}
\usage{
data(data.segment)
}
\description{
segment dataset
}
\examples{
data(data.segment)
}

