% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{data.stock}
\alias{data.stock}
\title{stock dataset}
\format{A data frame with 950 observations, 5 attributes and 5 labels}
\source{
UCI respository and Statlog collection
}
\usage{
data(data.stock)
}
\description{
stock dataset
}
\examples{
data(data.stock)
}

