% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{data.vehicle}
\alias{data.vehicle}
\title{vehicle dataset}
\format{A data frame with 846 observations, 18 attributes and 4 labels}
\source{
UCI respository and Statlog collection
}
\usage{
data(data.vehicle)
}
\description{
vehicle dataset
}
\examples{
data(data.vehicle)
}

