% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{data.vowel}
\alias{data.vowel}
\title{vowel dataset}
\format{A data frame with 528 observations, 10 attributes and 11 labels}
\source{
UCI respository and Statlog collection
}
\usage{
data(data.vowel)
}
\description{
vowel dataset
}
\examples{
data(data.vowel)
}

