% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{data.wine}
\alias{data.wine}
\title{Iris dataset}
\format{A data frame with 150 observations, 13 attributes and 3 labels}
\source{
Fisher, R. A. (1936) The use of multiple measurements in taxonomic problems. Annals of Eugenics, 7, Part II, 179-188.
}
\usage{
data(data.wine)
}
\description{
This famous (Fisher's or Anderson's) iris data set gives the measurements in centimeters of the variables sepal length and width and petal length and width, respectively, for 50 flowers from each of 3 species of iris. The species are Iris setosa, versicolor, and virginica.
}
\examples{
data(data.wine)
}

