% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{data.wisconsin}
\alias{data.wisconsin}
\title{wisconsin dataset}
\format{A data frame with 194 observations, 16 attributes and 16labels}
\source{
UCI respository and Statlog collection
}
\usage{
data(data.wisconsin)
}
\description{
wisconsin dataset
}
\examples{
data(data.wisconsin)
}

