% Generated by roxygen2 (4.0.1): do not edit by hand
\name{est.PL.MM}
\alias{est.PL.MM}
\title{MM method for estimating Plackett-Luce model parameters}
\usage{
est.PL.MM(dset)
}
\arguments{
\item{dset}{ranking dataset}
}
\value{
Estimated parameters, log-likelihood and number of iterations
}
\description{
est.PL.MM uses MM (minorization-maximization) algorithms to fit the Plackett-Luce model.
The R code was modified from the original Matlab code (plackmm) which can be found at
http://www.stat.psu.edu/~dhunter/code/btmatlab/
}
\examples{
# complete rankings
# create 2 ranks and 3 labels
ranks = matrix(0,2,3) # rows:number of ranks, column:number of labels
ranks[1,] = c(1,2,3) # L1>L2>L3
ranks[2,] = c(3,1,2) # L3>L1>L2
est.PL.MM(ranks)

# incomplate rankings
# create 3 ranks and 3 labels
ranks = matrix(0,3,3) # rows:number of ranks, column:number of labels
ranks[1,] = c(1,2,3) # L1>L2>L3
ranks[2,] = c(3,1,2) # L3>L1>L2
ranks[3,] = c(2,1,0) # L2>L1
est.PL.MM(ranks)
}
\references{
Hunter, D.R. (2004). MM algorithms for generalized Bradley-Terry models. Ann. of Stats., 32, 384-406
Plackett, R. (1975). The analysis of permutations. Applied Stat., 24, 193-202.
}

