% Generated by roxygen2 (4.0.1): do not edit by hand
\name{est.PL2.NR}
\alias{est.PL2.NR}
\title{NR method for estimating Plackett-Luce (2-vases model) parameters}
\usage{
est.PL2.NR(dset, L)
}
\arguments{
\item{dset}{ranking dataset}

\item{L}{switch point between the first vase and second vase}
}
\value{
Estimated parameters, log-likelihood and number of iterations
}
\description{
uses Newton-Raphson algorithm to fit the Plackett-Luce with 2-vases model.
}

