% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eval.CV}
\alias{eval.CV}
\title{K-flods cross validation}
\usage{
eval.CV(file, method, k = 15, nFolds = 10, randomness = TRUE)
}
\arguments{
\item{file}{lable ranking dataset}

\item{method}{method to fit Plackett-Luce model}

\item{k}{number of k-neiborhoods}

\item{nFlods}{number of flods in the cross validation}

\item{randomness}{logical indicating whether}
}
\value{
Kendall's tau, Spearman's rank and log-likelihood
}
\description{
K-flods cross validation
}
\examples{
exp.CV("iris", method="plmm", k=15, nFlods=10, randomness=TRUE)
}
\references{
ref1
ref2
ref3
}

