% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eval.getConcordantPairs}
\alias{eval.getConcordantPairs}
\title{Number of concordant pairs (Kendall's tau)}
\usage{
eval.getConcordantPairs(t)
}
\arguments{
\item{t}{m*2 matrix where the first column is the actual rank
and the second column is the predicted rank}
}
\value{
The number of concordant pairs
}
\description{
eval.getConcordantPairs returns the number of concordant pairs referred to
Kendall's tau correlation coefficient definition
}
\examples{
actula_rank = c(1,2,3)     # L1>L2>L3
predicted_rank = c(1,3,2)  # L1>L3>L2
t <- data.frame(actula_rank, predicted_rank)
eval.getConcordantPairs(t)
}

