% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eval.KendallTau}
\alias{eval.KendallTau}
\title{Kendall's tau rank correlation coefficient}
\usage{
eval.KendallTau(t)
}
\arguments{
\item{t}{m*2 matrix where
: 1st column is the actual rank
: 2nd column is the predicted rank}
}
\value{
Kendall's tau rank correlation coefficient which is based on this fomular:
  C-D / (1/2)n(n-1) where
  C is number of concordant pairs
  D is number of disconcordant pairs
  n is number of labels to rank
}
\description{
eval.KendallTau returns Kendall's tau rank correlation coefficient.
}
\examples{
actula_rank = c(1,2,3)     # L1>L2>L3
predicted_rank = c(1,3,2)  # L1>L3>L2
t <- data.frame(actula_rank, predicted_rank)
eval.KendallTau(t)
}

