% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eval.LRatioTest}
\alias{eval.LRatioTest}
\title{Likelihood ratio test}
\usage{
eval.LRatioTest(altLogL, nullLogL, dof, alpha = 0.05)
}
\arguments{
\item{altLogL}{loglikelihood of an alternative model.}

\item{nullLogL}{loglikelihood of a null model.}

\item{dof}{degree of freedoms. dof = m-1 where m is the number of labels to rank}

\item{alpha}{significance level (default=0.05).}
}
\value{
A logical value (h-value) with the rejection decision from conducting a likelihood ratio test of model specification.
}
\description{
eval.LRatioTest performs a statistical test using the loglikelihood objective function
to compare the fits of two models.
}
\examples{
altLogL = c(-10.87)   # loglikelihood from an alternative model
nullLogL = c(-10.039) # loglikelihood from a hull model
dof = 2 # degrees of freedom
eval.LRatioTest(altLogL, nullLogL, dof)
}

