% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eval.MAE}
\alias{eval.MAE}
\title{Mean absolute error performance}
\usage{
eval.MAE(actual, predicted)
}
\arguments{
\item{actual}{a numeric vector of actual parameters.}

\item{predicted}{a numeric vector of predicted parameters.}
}
\value{
The mean absolute error.
}
\description{
eval.MAE measures the performace as the mean of absolute errors of the actual parameters and predicted parameters.
}
\examples{
actual <- c(1,2,1)
predicted <- c(1,1,1)
eval.MAE(actual,predicted)
}

