% Generated by roxygen2 (4.0.1): do not edit by hand
\name{eval.SpearmanRho}
\alias{eval.SpearmanRho}
\title{Spearman's rank correlation coefficient}
\usage{
eval.SpearmanRho(t)
}
\arguments{
\item{t}{m*2 matrix where
: 1st column is the actual rank
: 2nd column is the predicted rank}
}
\value{
Spearman's rho which is based on this fomular:
1 - ( (6*sum(d^2))/n(n^2-1) ) , where d is a displacement of each element in the rank
}
\description{
eval.SpearmanRho returns the Spearman's rank correlation coefficient or Spearman's rho
}
\examples{
actula_rank = c(1,2,3)     # L1>L2>L3
predicted_rank = c(1,3,2)  # L1>L3>L2
t <- data.frame(actula_rank, predicted_rank)
eval.SpearmanRho(t)
}

