% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genRank1v}
\alias{genRank1v}
\title{Generate artificial dataset from single vase model distribution}
\usage{
genRank1v(nLabels, nObs)
}
\arguments{
\item{nLabels}{number of labels to rank}

\item{nObs}{number of observations}
}
\value{
Probability distribution and ranks
}
\description{
This is useful for performing inferrence tasks for ranking model
}
\examples{
## Create artificial dataset with 4 labels to rank and 10 instances
lables = 4
observations = 10
dataset = genRank1v(lables, observations)
}

