% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genRank2v}
\alias{genRank2v}
\title{Generate artificial dataset from two vases model distribution}
\usage{
genRank2v(nLabels, nObs, L)
}
\arguments{
\item{nLabels}{number of labels to rank}

\item{nObs}{number of observations}

\item{L}{lag or switch position between 1st vase and 2nd vase}
}
\value{
Probability distribution for the 1st vase, 2nd vase and ranks
}
\description{
This is useful for performing inferrence tasks for ranking model
}
\examples{
## Create artificial dataset with 4 labels to rank and 10 instances
lables = 4
observations = 10
L = 2
dataset = genRank2v(lables, observations, L)
}

