% Generated by roxygen2 (4.0.1): do not edit by hand
\name{knn.getRanks}
\alias{knn.getRanks}
\title{Returns complete ranks associated to the nearest neighbors}
\usage{
knn.getRanks(dset, instance)
}
\arguments{
\item{data}{}

\item{instance}{}
}
\description{
Returns complete ranks associated to the nearest neighbors
}

