% Generated by roxygen2 (4.0.1): do not edit by hand
\name{knn.search}
\alias{knn.search}
\title{Find the nearest neighbors}
\usage{
knn.search(X, Y, k = "10", algo = "kd_tree")
}
\arguments{
\item{data}{an input data matrix.}

\item{query}{a query data matrix}

\item{k}{the number of nearest neighbors to search. The default value is set to 10}

\item{algo}{nearest neightbor seaching algorithms. The default value is kd-tree}
}
\value{
???
}
\description{
Finds the nearest neighbor in X for each point in Y.
X is an mx-by-n matrix and Y is an my-by-n matrix.
Rows of X and Y correspond to observations and columns correspond to variables.
}

